+++ src/main/kotlin/com/duolingo/clone/language_backend/dto/UserProfileResponse.kt
@@ -7,7 +7,7 @@ data class UserProfileResponse(
     val fullName: String,
     val username: String,
     val joinedAt: Instant?,
-    val totalXp: Long,
+    val totalXp: Int,
     val currentStreak: Int,
     val lingots: Int,
     val heartsCount: Int,
